backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
google-generativeai>=0.3.0
aiofiles>=23.2.0
httpx>=0.26.0
openai>=1.0.0
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0